Inventory tables check if they have the properties they need
