+++ plugins/main/public/components/agents/syscollector/inventory.tsx
@@ -11,28 +11,18 @@
  */
 
 import React from 'react';
-import { EuiFlexGroup, EuiFlexItem, EuiCallOut, EuiPanel } from '@elastic/eui';
+import { EuiFlexGroup, EuiFlexItem, EuiCallOut } from '@elastic/eui';
 import { InventoryMetrics } from './components/syscollector-metrics';
-import {
-  netaddrColumns,
-  netifaceColumns,
-  processColumns,
-  portsColumns,
-  packagesColumns,
-  windowsUpdatesColumns,
-} from './columns';
-import {
-  API_NAME_AGENT_STATUS,
-  SEARCH_BAR_WQL_VALUE_SUGGESTIONS_COUNT,
-} from '../../../../common/constants';
-import { TableWzAPI } from '../../common/tables';
-import { WzRequest } from '../../../react-services';
-import { get as getLodash } from 'lodash';
+import { API_NAME_AGENT_STATUS } from '../../../../common/constants';
 import { compose } from 'redux';
 import { withGuard } from '../../common/hocs';
 import { PromptAgentNeverConnected } from '../prompts';
-
-const sortFieldSuggestion = (a, b) => (a.label > b.label ? 1 : -1);
+import NetworkInterfacesTable from './components/network-interfaces-table';
+import NetworkPortsTable from './components/network-ports-table';
+import NetworkSettingsTable from './components/network-settings-table';
+import WindowsUpdatesTable from './components/windows-updates-table';
+import ProcessesTable from './components/processes-table';
+import PackagesTable from './components/packages-table';
 
 export const SyscollectorInventory = compose(
   withGuard(
@@ -72,359 +62,33 @@ export const SyscollectorInventory = compose(
 
       <EuiFlexGroup gutterSize='s'>
         <EuiFlexItem grow={2} style={{ marginRight: 4, marginTop: 0 }}>
-          <EuiPanel paddingSize='m' style={{ margin: '12px 16px 12px 16px' }}>
-            <TableWzAPI
-              title='Network interfaces'
-              tableColumns={netifaceColumns}
-              tableInitialSortingField={netifaceColumns[0].field}
-              endpoint={`/syscollector/${
-                agent.id
-              }/netiface?select=${netifaceColumns
-                .map(({ field }) => field)
-                .join(',')}`}
-              searchTable
-              downloadCsv
-              showReload
-              tablePageSizeOptions={[10, 25, 50, 100]}
-              searchBarWQL={{
-                suggestions: {
-                  field(currentValue) {
-                    return netifaceColumns
-                      .map(item => ({
-                        label: item.field,
-                        description: `filter by ${item.name}`,
-                      }))
-                      .sort(sortFieldSuggestion);
-                  },
-                  value: async (currentValue, { field }) => {
-                    try {
-                      const response = await WzRequest.apiReq(
-                        'GET',
-                        `/syscollector/${agent.id}/netiface`,
-                        {
-                          params: {
-                            distinct: true,
-                            limit: SEARCH_BAR_WQL_VALUE_SUGGESTIONS_COUNT,
-                            select: field,
-                            sort: `+${field}`,
-                            ...(currentValue
-                              ? { q: `${field}~${currentValue}` }
-                              : {}),
-                          },
-                        },
-                      );
-                      return response?.data?.data.affected_items.map(item => ({
-                        label: getLodash(item, field),
-                      }));
-                    } catch (error) {
-                      return [];
-                    }
-                  },
-                },
-              }}
-              tableProps={{
-                tableLayout: 'auto',
-              }}
-            />
-          </EuiPanel>
+          <NetworkInterfacesTable agent={agent} />
         </EuiFlexItem>
         <EuiFlexItem grow={2} style={{ marginLeft: 4, marginTop: 0 }}>
-          <EuiPanel paddingSize='m' style={{ margin: '12px 16px 12px 16px' }}>
-            <TableWzAPI
-              title='Network ports'
-              tableColumns={portsColumns[soPlatform]}
-              tableInitialSortingField={portsColumns[soPlatform][0].field}
-              endpoint={`/syscollector/${agent.id}/ports?select=${portsColumns[
-                soPlatform
-              ]
-                .map(({ field }) => field)
-                .join(',')}`}
-              searchTable
-              downloadCsv
-              showReload
-              tablePageSizeOptions={[10, 25, 50, 100]}
-              searchBarWQL={{
-                suggestions: {
-                  field(currentValue) {
-                    return portsColumns[soPlatform]
-                      .map(item => ({
-                        label: item.field,
-                        description: `filter by ${item.name}`,
-                      }))
-                      .sort(sortFieldSuggestion);
-                  },
-                  value: async (currentValue, { field }) => {
-                    try {
-                      const response = await WzRequest.apiReq(
-                        'GET',
-                        `/syscollector/${agent.id}/ports`,
-                        {
-                          params: {
-                            distinct: true,
-                            limit: SEARCH_BAR_WQL_VALUE_SUGGESTIONS_COUNT,
-                            select: field,
-                            sort: `+${field}`,
-                            ...(currentValue
-                              ? { q: `${field}~${currentValue}` }
-                              : {}),
-                          },
-                        },
-                      );
-                      return response?.data?.data.affected_items.map(item => ({
-                        label: getLodash(item, field),
-                      }));
-                    } catch (error) {
-                      return [];
-                    }
-                  },
-                },
-              }}
-              tableProps={{
-                tableLayout: 'auto',
-              }}
-            />
-          </EuiPanel>
+          <NetworkPortsTable agent={agent} soPlatform={soPlatform} />
         </EuiFlexItem>
       </EuiFlexGroup>
 
       <EuiFlexGroup gutterSize='s'>
         <EuiFlexItem grow={3} style={{ marginRight: 4 }}>
-          <EuiPanel paddingSize='m' style={{ margin: '12px 16px 12px 16px' }}>
-            <TableWzAPI
-              title='Network settings'
-              tableColumns={netaddrColumns}
-              tableInitialSortingField={netaddrColumns[0].field}
-              endpoint={`/syscollector/${
-                agent.id
-              }/netaddr?select=${netaddrColumns
-                .map(({ field }) => field)
-                .join(',')}`}
-              searchTable
-              downloadCsv
-              showReload
-              tablePageSizeOptions={[10, 25, 50, 100]}
-              searchBarWQL={{
-                suggestions: {
-                  field(currentValue) {
-                    return netaddrColumns
-                      .map(item => ({
-                        label: item.field,
-                        description: `filter by ${item.name}`,
-                      }))
-                      .sort(sortFieldSuggestion);
-                  },
-                  value: async (currentValue, { field }) => {
-                    try {
-                      const response = await WzRequest.apiReq(
-                        'GET',
-                        `/syscollector/${agent.id}/netaddr`,
-                        {
-                          params: {
-                            distinct: true,
-                            limit: SEARCH_BAR_WQL_VALUE_SUGGESTIONS_COUNT,
-                            select: field,
-                            sort: `+${field}`,
-                            ...(currentValue
-                              ? { q: `${field}~${currentValue}` }
-                              : {}),
-                          },
-                        },
-                      );
-                      return response?.data?.data.affected_items.map(item => ({
-                        label: getLodash(item, field),
-                      }));
-                    } catch (error) {
-                      return [];
-                    }
-                  },
-                },
-              }}
-              tableProps={{
-                tableLayout: 'auto',
-              }}
-            />
-          </EuiPanel>
+          <NetworkSettingsTable agent={agent} />
         </EuiFlexItem>
         {agent && agent.os && agent.os.platform === 'windows' && (
           <EuiFlexItem grow={1} style={{ marginLeft: 4 }}>
-            <EuiPanel paddingSize='m' style={{ margin: '12px 16px 12px 16px' }}>
-              <TableWzAPI
-                title='Windows updates'
-                tableColumns={windowsUpdatesColumns}
-                tableInitialSortingField={windowsUpdatesColumns[0].field}
-                endpoint={`/syscollector/${
-                  agent.id
-                }/hotfixes?select=${windowsUpdatesColumns
-                  .map(({ field }) => field)
-                  .join(',')}`}
-                searchTable
-                downloadCsv
-                showReload
-                tablePageSizeOptions={[10, 25, 50, 100]}
-                searchBarWQL={{
-                  suggestions: {
-                    field(currentValue) {
-                      return windowsUpdatesColumns
-                        .map(item => ({
-                          label: item.field,
-                          description: `filter by ${item.name}`,
-                        }))
-                        .sort(sortFieldSuggestion);
-                    },
-                    value: async (currentValue, { field }) => {
-                      try {
-                        const response = await WzRequest.apiReq(
-                          'GET',
-                          `/syscollector/${agent.id}/hotfixes`,
-                          {
-                            params: {
-                              distinct: true,
-                              limit: SEARCH_BAR_WQL_VALUE_SUGGESTIONS_COUNT,
-                              select: field,
-                              sort: `+${field}`,
-                              ...(currentValue
-                                ? { q: `${field}~${currentValue}` }
-                                : {}),
-                            },
-                          },
-                        );
-                        return response?.data?.data.affected_items.map(
-                          item => ({
-                            label: getLodash(item, field),
-                          }),
-                        );
-                      } catch (error) {
-                        return [];
-                      }
-                    },
-                  },
-                }}
-                tableProps={{
-                  tableLayout: 'auto',
-                }}
-              />
-            </EuiPanel>
+            <WindowsUpdatesTable agent={agent} />
           </EuiFlexItem>
         )}
       </EuiFlexGroup>
 
       <EuiFlexGroup gutterSize='s'>
         <EuiFlexItem>
-          <EuiPanel paddingSize='m' style={{ margin: '12px 16px 12px 16px' }}>
-            <TableWzAPI
-              title='Packages'
-              tableColumns={packagesColumns[soPlatform]}
-              tableInitialSortingField={packagesColumns[soPlatform][0].field}
-              endpoint={`/syscollector/${
-                agent.id
-              }/packages?select=${packagesColumns[soPlatform]
-                .map(({ field }) => field)
-                .join(',')}`}
-              searchTable
-              downloadCsv
-              showReload
-              tablePageSizeOptions={[10, 25, 50, 100]}
-              searchBarWQL={{
-                suggestions: {
-                  field(currentValue) {
-                    return packagesColumns[soPlatform]
-                      .map(item => ({
-                        label: item.field,
-                        description: `filter by ${item.name}`,
-                      }))
-                      .sort(sortFieldSuggestion);
-                  },
-                  value: async (currentValue, { field }) => {
-                    try {
-                      const response = await WzRequest.apiReq(
-                        'GET',
-                        `/syscollector/${agent.id}/packages`,
-                        {
-                          params: {
-                            distinct: true,
-                            limit: SEARCH_BAR_WQL_VALUE_SUGGESTIONS_COUNT,
-                            select: field,
-                            sort: `+${field}`,
-                            ...(currentValue
-                              ? { q: `${field}~${currentValue}` }
-                              : {}),
-                          },
-                        },
-                      );
-                      return response?.data?.data.affected_items.map(item => ({
-                        label: getLodash(item, field),
-                      }));
-                    } catch (error) {
-                      return [];
-                    }
-                  },
-                },
-              }}
-              tableProps={{
-                tableLayout: 'auto',
-              }}
-            />
-          </EuiPanel>
+          <PackagesTable agent={agent} soPlatform={soPlatform} />
         </EuiFlexItem>
       </EuiFlexGroup>
 
       <EuiFlexGroup gutterSize='s'>
         <EuiFlexItem>
-          <EuiPanel paddingSize='m' style={{ margin: '12px 16px 12px 16px' }}>
-            <TableWzAPI
-              title='Processes'
-              tableColumns={processColumns[soPlatform]}
-              tableInitialSortingField={processColumns[soPlatform][0].field}
-              endpoint={`/syscollector/${
-                agent.id
-              }/processes?select=${processColumns[soPlatform]
-                .map(({ field }) => field)
-                .join(',')}`}
-              searchTable
-              downloadCsv
-              showReload
-              tablePageSizeOptions={[10, 25, 50, 100]}
-              searchBarWQL={{
-                suggestions: {
-                  field(currentValue) {
-                    return processColumns[soPlatform]
-                      .map(item => ({
-                        label: item.field,
-                        description: `filter by ${item.name}`,
-                      }))
-                      .sort(sortFieldSuggestion);
-                  },
-                  value: async (currentValue, { field }) => {
-                    try {
-                      const response = await WzRequest.apiReq(
-                        'GET',
-                        `/syscollector/${agent.id}/processes`,
-                        {
-                          params: {
-                            distinct: true,
-                            limit: SEARCH_BAR_WQL_VALUE_SUGGESTIONS_COUNT,
-                            select: field,
-                            sort: `+${field}`,
-                            ...(currentValue
-                              ? { q: `${field}~${currentValue}` }
-                              : {}),
-                          },
-                        },
-                      );
-                      return response?.data?.data.affected_items.map(item => ({
-                        label: getLodash(item, field),
-                      }));
-                    } catch (error) {
-                      return [];
-                    }
-                  },
-                },
-              }}
-              tableProps={{
-                tableLayout: 'auto',
-              }}
-            />
-          </EuiPanel>
+          <ProcessesTable agent={agent} soPlatform={soPlatform} />
         </EuiFlexItem>
       </EuiFlexGroup>
     </div>
